+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/plugin/factory/StreamStatePluginFactory.kt
@@ -170,7 +170,7 @@ public class StreamStatePluginFactory(
             stateRegistry = stateRegistry,
             syncManager = syncManager,
             eventHandler = eventHandler,
-            globalState = mutableGlobalState,
+            mutableGlobalState = mutableGlobalState,
             queryingChannelsFree = isQueryingFree,
             statePluginConfig = config,
         )