+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `DraftMessage` entity to represent a draft message. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
 
 ### ⚠️ Changed
 
@@ -16,6 +17,9 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `ChatClient.createDraftMessage()` method to create a draft message. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
+- Add `ChatClient.deleteDraftMessages()` method to delete a draft message. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
+- Add `ChatClient.queryDraftMessages()` method to query draft messages. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
 
 ### ⚠️ Changed
 
@@ -38,6 +42,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `GlobalState.channelDraftMessages` and `GlobalState.threadDraftMessages` properties providing access to the draft messages. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682) 
 
 ### ⚠️ Changed